Update patch set 4
Patch Set 4:

It appears there is some misunderstanding here – the VOs are not exactly the same as CDS objects. VOs have some more data elements that is not yet part of Boreas and hence not yet in CDS. For example a Project VO has not only Pipeline and Notebooks in it but also other objects such as Topic, Publisher, Subscriber, etc., which will be null now but part of the final Project structure. VOs were designed as part of the target state integration. So we do not have to enhance the VOs with each Acumos Release.

Patch-set: 4