Update patch set 4
Patch Set 4: Code-Review-1

Running independent services is what we wanted and is the best design, which is good architecture. 

However, “do not call each other” is not acceptable, what is the purpose of encapsulating the individual services? if they do not talk to each other and keep duplicating the backend calls in all the services. In this specific code review, would you have projectExists() method on both ProjectService and NotebookService? Ideally, it should be in ProjectService not in NotebookService. The Notebook or Pipeline services must call the ProjectService endpoint to verify if the project exists.

I clarified the concern I see and proposed the change we should do.

Please justify, if you do not agree with above suggestion. If you agree, then please move the method “projectExists()” from NotebookService to ProjectService. Thanks.

Patch-set: 4
Reviewer: Gerrit User 3 <3@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=-1