Update patch set 28
Patch Set 28: Code-Review-1

(3 comments)

Thanks for the changes and the time spent on this. This is all great work!! Posted minor comments/concerns on dashboard/home-webcomponent modules. Soon, will post on remaining modules.

Please address all the comments and make sure some are not ignored, as it will be hard for reviewers to track. 
If addressed in a patch, please specify which comments are addressed and which ones will be addressed in the following cpatches. 
If not planning to address any comments, then please provide the technical reasons to make such choice.

Also, while addressing comments, please address them for all modules as they are applicable as reviewer may not repeat the same comment on all the module files.

Patch-set: 28
Label: Code-Review=-1