Update patch set 8
Patch Set 8:

> Patch Set 8: Code-Review-1
> 
> (1 comment)

Using "!important" in this case cannot be avoided due to the way that older versions of Firefox implement web components. For example, only fonts and color styles should propagate to web components, but in older versions of Firefox other styles also propagate as well.

This is a workaround for this in order to properly support older versions of Firefox.

Patch-set: 8