Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> > Patch Set 8: Code-Review-1
> > 
> > (1 comment)
> 
> Using "!important" in this case cannot be avoided due to the way that older versions of Firefox implement web components. For example, only fonts and color styles should propagate to web components, but in older versions of Firefox other styles also propagate as well.
> 
> This is a workaround for this in order to properly support older versions of Firefox.


I assume we provided this temporary workaround to support testing team, who has older version firefox and may not be allowed to install latest versions or chrome browser because of firewall rules in the company. May be it is OK for now. But, we must avoid the "!important" usage.

Firefox version 63 seems to be happy with web components and version 63 is released in Oct 2018. 

It should be ok to say Acumos UI support Firefox version 63 and latest, as the version 63 is already 6 months old. 

Now-a-days 6 month old version browser is potentially vulnerable to hacking, so most users are at least up-to-date to last 1 to 3 month old versions. 

So, we must avoid these hacks or temporary workaround which will impact the CSS maintenance in long run and say that we support Firefox version 63 and latest. 

Please confirm, how and when you want to remove these. Is it in this patch? If not, please create a followup Jira item and post it here, so that we can track it. Thanks.

Patch-set: 8