Update patch set 3
Patch Set 3:

> You can release multiple containers from the same project at once,
 > but all of them must be listed in one file.  The wiki page gives an
 > example.

Yes, you are correct, but the "ref" is not same for all of them.  Also release job fails if the specified tag already exist, so we tried adding text in the tag value (but even that's also not accepted by release job).  We have same version for all MLWB UI components, but the "ref" id is not same for all of them and that's the issue.

Patch-set: 3
Reviewer: Gerrit User 11 <11@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=0