Update patch set 1
Patch Set 1:

I think that what happened here was the your "remerge" comment triggered the merge job to run on the commit that was at the tip of master when you posted your comment. Maybe you expected it to checkout the result of the merge?

Also I see that you submitted this to the clio branch.  That's kind of weird, you should really keep all the release files on master branch just to keep track.

Patch-set: 1